requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==4.1.9
django-cors-headers==4.4.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.6.1
mongoengine==0.28.2
pymongo==3.12.3
pytz==2024.1
sqlparse==0.2.4
tzdata==2024.1